A bulky, but rather brainless pass at treating the important strings in struct waypoint as a class instead of C string.  Of course, that struct has one member - a C string - but this starts to hightlight the places that are going to be problem areas for us.
With NEW_STRINGS turned on, we basically run, but suffer from a double
encoding of non-ascii strings.  Every format seems to be affected, so
it's probably in CET somewhere.  This isn't a QString thing in any obvious
way, as the underlying data types are still actually C Strings.

From here, additional passes can be made, either turning the individual
string pointer conversions off and whacking on a few files at a time or 
going full money and making String a QString.

It does highlight the need for better vararg-style functions like warning,
fatal, and our debugging printfs to to know about stream operators.